+++ eulxml/xmlmap/eadmap.py
@@ -45,7 +45,7 @@ class _EadBase(xmlmap.XmlObject):
 class Note(_EadBase):
     """EAD note."""
     ROOT_NAME = 'note'
-    content = xmlmap.StringListField("e:p")
+    content = xmlmap.NodeListField("e:p", xmlmap.XmlObject)   # ?? (to allow formatting)
     "list of paragraphs - `p`"
 
 
@@ -53,7 +53,7 @@ class Section(_EadBase):
     """Generic EAD section.  Currently only has mappings for head, paragraph, and note."""
     head = xmlmap.NodeField("e:head", xmlmap.XmlObject)
     "heading - `head`"
-    content = xmlmap.NodeListField("e:p", xmlmap.XmlObject)       # ??
+    content = xmlmap.NodeListField("e:p", xmlmap.XmlObject)   # ?? (to allow formatting)
     "list of paragraphs - `p`"
     note = xmlmap.NodeField("e:note", Note)
     ":class:`Note`"
@@ -188,6 +188,8 @@ class DigitalArchivalObject(_EadBase):
     'title'
     href = xmlmap.StringField("@xlink:href")
     'url where the digital archival object can be accessed'
+    show = xmlmap.StringField("@xlink:show")
+    'attribute to determine how the resource should be displayed'
 
 
 class DescriptiveIdentification(_EadBase):
@@ -246,6 +248,8 @@ class Component(_EadBase):
     "bibliography :class:`Section` - `bibliograhy`"
     scope_content = xmlmap.NodeField("e:scopecontent", Section)
     "scope and content :class:`Section` - `scopecontent`"
+    process_info = xmlmap.NodeField("e:processinfo", Section)
+    "processing infomration :class:`Section` - `processinfo`"
     arrangement = xmlmap.NodeField("e:arrangement", Section)
     "arrangement :class:`Section` - `arrangement`"
     other = xmlmap.NodeField("e:otherfindaid", Section)
@@ -349,6 +353,8 @@ class Index(Section):
     entry = xmlmap.NodeListField("e:indexentry", IndexEntry)
     "list of :class:`IndexEntry` - `indexentry`; entry in the index"
     id = xmlmap.StringField("@id")
+    note = xmlmap.NodeField("e:note", Note)
+    ":class:`Note`"
 
 
 class ArchivalDescription(_EadBase):
@@ -393,6 +399,8 @@ class ArchivalDescription(_EadBase):
     "bibliography :class:`Section` - `bibliograhy`"
     scope_content = xmlmap.NodeField("e:scopecontent", Section)
     "scope and content :class:`Section` - `scopecontent`"
+    process_info = xmlmap.NodeField("e:archdesc/e:processinfo", Section)
+    "processing information :class:`Section` - `processinfo`"
     arrangement = xmlmap.NodeField("e:arrangement", Section)
     "arrangement :class:`Section` - `arrangement`"
     other = xmlmap.NodeField("e:otherfindaid", Section)